+++ tiled/server/utils.py
@@ -9,6 +9,7 @@ from starlette.types import Scope
 from ..access_control.access_policies import NO_ACCESS
 from ..access_control.protocols import AccessPolicy
 from ..adapters.mapping import MapAdapter
+from ..queries import AccessTagsFilter
 from ..server.schemas import Principal
 from ..type_aliases import AccessTags, Scopes
 
@@ -114,5 +115,9 @@ async def filter_for_access(
                 entry = EMPTY_NODE
             else:
                 for query in queries:
+                    if isinstance(query, AccessTagsFilter) and hasattr(
+                        entry, "resolve_access_tag_ids"
+                    ):
+                        query.tag_ids = await entry.resolve_access_tag_ids(query.tags)
                     entry = entry.search(query)
     return entry